+++ pyhpeimc/tests/test_machine.py
