+++ pyhpeimc/tests/test_pyhpeimc_plat_device.py
@@ -5,9 +5,8 @@ This module is used for testing the functions within the pyhpeimc.plat.device mo
 """
 
 from unittest import TestCase
-
 from nose.plugins.skip import SkipTest
-
+from pyhpeimc.tests.test_machine import *
 from pyhpeimc.plat.device import *
