+++ pyhpeimc/tests/test_pyhpeimc_plat_operator.py
@@ -4,10 +4,10 @@ This module is used for testing the functions within the pyhpeimc.plat.operator 
 
 """
 
+
 from unittest import TestCase
-
+from pyhpeimc.tests.test_machine import *
 from pyhpeimc.plat.operator import *
-
 
 class TestCreate_operator(TestCase):
     def test_create_operator(self):
