+++ pyhpeimc/tests/test_pyhpeimc_plat_perf.py
@@ -4,10 +4,10 @@ This module is used for testing the functions within the pyhpeimc.plat.perf modu
 
 """
 
+
 from unittest import TestCase
-
+from pyhpeimc.tests.test_machine import *
 from pyhpeimc.plat.perf import *
-
 
 class TestAdd_perf_task(TestCase):
     def test_add_perf_task(self):
