+++ pyhpeimc/tests/test_pyhpeimc_plat_system.py
@@ -4,10 +4,10 @@ This module is used for testing the functions within the pyhpeimc.plat.system mo
 
 """
 
-from unittest import TestCase
 
+from unittest import TestCase
+from pyhpeimc.tests.test_machine import auth
 from pyhpeimc.plat.system import *
-from tests import auth
 
 
 class TestGet_system_vendors(TestCase):
