+++ pyhpeimc/tests/test_pyhpeimc_plat_vrm.py
@@ -4,11 +4,12 @@ This module is used for testing the functions within the pyhpeimc.plat.vrm modul
 
 """
 
+
 from unittest import TestCase
-
 from nose.plugins.skip import SkipTest
-
+from pyhpeimc.tests.test_machine import *
 from pyhpeimc.plat.vrm import *
+
 
 
 #####Test get_vm_host_info for Multiple Vendor Devices
