+++ pyhpeimc/tests/test_pyhpeimc_wsm_clientinfo.py
@@ -4,9 +4,11 @@ This module is used for testing the functions within the pyhpeimc.wsm.clientinfo
 
 """
 
-from unittest import TestCase
 
+from unittest import TestCase
+from pyhpeimc.tests.test_machine import *
 from pyhpeimc.wsm.clientinfo import *
+
 
 
 class Test_Get_client_info_all(TestCase):
